Only partially sort the first k nearest neighbors instead of a full sort.
Doing a knn-query works as follows:

- for every rtree: query its k nearest neighbors

- for all knn query results: pick the k nearest

therefore, as we only need the k nearest, a partial sort suffices.

That is, instead of fully sorting k * number of rtree results,
we partially sort only for the first k using `nth_element`.

This makes sure the range is correctly sorted, but only up to the
first n elements (our k nearest neighbors), speeding up the query.